ping-viewer-next-frontend: Split and separate main application resources loading from widgets route"
